refactor/product-delete: 상품 삭제 어드민 권한 추가

+++ src/main/java/org/example/eatopia/domain/product/service/command/ProductCommandServiceImpl.java
@@ -96,8 +96,9 @@ public class ProductCommandServiceImpl implements ProductCommandService {
     public void deleteProduct(Long productId, Long userId) {
 
         Product product = productQueryService.getProductOrElseThrow(productId);
+        User user = userQueryService.getUserEntityById(userId);
 
-        product.verifySeller(userId);
+        product.verifySellerOrAdmin(userId, user.isAdmin());
 
         productRepository.delete(product);
     }
